hdrhistogram

Module iterators

Module containing the implementations of all Histogram iterators.

Modules

all

An iterator that iterates over histogram values.

linear

An iterator that iterates linearly over histogram values.

log

An iterator that iterates logarithmically over histogram values.

quantile

An iterator that iterates over histogram quantiles.

recorded

An iterator that iterates over recorded histogram values.

Structs

HistogramIterator

HistogramIterator provides a base iterator for a Histogram.

IterationValue

The value emitted at each step when iterating over a Histogram.

PickMetadata

Extra information about the picked point in the histogram provided by the picker.

Traits

PickyIterator

A trait for designing an subset iterator over values in a Histogram.