hdrhistogram::sync

Struct SyncHistogram

pub struct SyncHistogram<C: Counter> { /* private fields */ }

A Histogram that can be written to by multiple threads concurrently.

Each writer thread should have a Recorder, which allows it to record new samples without synchronization. New recorded samples are made available through this histogram by calling SyncHistogram::refresh, which blocks until it has synchronized with every recorder.

Implementations

impl<C: Counter> SyncHistogram<C>

pub fn refresh(&mut self)

Block until writes from all Recorder instances for this histogram have been incorporated.

pub fn refresh_timeout(&mut self, timeout: Duration)

Block until writes from all Recorder instances for this histogram have been incorporated, or until the given amount of time has passed.

pub fn recorder(&self) -> Recorder<C>

Obtain another multi-threaded writer for this histogram.

Note that writes made to the Recorder will not be visible until the next call to SyncHistogram::refresh.