Crate hdrhistogram

Rust binding for HdrHistogram_c

This crate implements bindings for HdrHistogram_c, a flexible library for recording histograms without having to know very much about the data being histogrammed.

The top-level type is Histogram.

Example

This sets up a histogram to record values in the range 1..1000_000 with 2 significant figures of precision. It then records

let mut h = Histogram::init(1, 1000000, 2).unwrap();
 
h.record_value(1);
h.record_value(10);
h.record_values(100, 40);

assert_eq!(h.total_count(), 42);
assert_eq!(h.min(), 1);
assert_eq!(h.max(), 100);

Structs

CountIterItem

F64CountIterItem

F64Histogram

F64LinearIter

F64LogIter

F64PercentileIter

F64PercentileIterItem

F64RecordedIter

HistoErr
Catch-all error return.

Histogram

LinearIter

LogIter

PercentileIter

PercentileIterItem

RecordedIter