hdrhistogram

Trait Counter

pub trait Counter:
    Num
    + ToPrimitive
    + FromPrimitive
    + Saturating
    + CheckedSub
    + CheckedAdd
    + Copy
    + PartialOrd<Self>
    + Debug {
    // Required methods
    fn as_f64(&self) -> f64;
    fn as_u64(&self) -> u64;
}

This trait represents the operations a histogram must be able to perform on the underlying counter type. The ToPrimitive trait is needed to perform floating point operations on the counts (usually for quantiles). The FromPrimitive to convert back into an integer count. Partial ordering is used for threshholding, also usually in the context of quantiles.

Required Methods

fn as_f64(&self) -> f64

Counter as a f64.

fn as_u64(&self) -> u64

Counter as a u64.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.